requirements.txt
streamlit
pandas
requests
beautifulsoup4
python-dateutil
matplotlib
transformers==4.56.2
torch==2.5.1